Add files via upload
This project is about Encryption messages from form by Caesar Chipper algorithm with encrypt key and hiding or decoding encryption messages to an image by the Least Significant Bit method, and then the result of decoding be called Stego Image. Stego Image was extracted or decoded to be encryption messages and decrypted to be the original message or original form using the same encrypted key.